# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Recruit API",
        "description": "採用広報サイト + 応募管理ダッシュボード API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/admin/media": {
            "get": {
                "tags": [
                    "Admin/Media"
                ],
                "summary": "メディア一覧",
                "operationId": "e858e24ba20eb59ee7f6e6699cc097f3",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "メディア一覧"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin/Media"
                ],
                "summary": "メディアアップロード",
                "operationId": "3517df0a72986953693aad59d24b5a2d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "アップロード成功"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/media/{media}": {
            "delete": {
                "tags": [
                    "Admin/Media"
                ],
                "summary": "メディア削除",
                "operationId": "4393617f7d25f60975b7e970e1158725",
                "parameters": [
                    {
                        "name": "media",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "削除成功"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/members/{slug}": {
            "get": {
                "tags": [
                    "Members"
                ],
                "summary": "社員詳細（公開）",
                "operationId": "c8086787161aa806b9fb6b009c41e449",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "社員詳細"
                    },
                    "404": {
                        "description": "見つからない"
                    }
                }
            }
        },
        "/admin/members": {
            "get": {
                "tags": [
                    "Admin/Members"
                ],
                "summary": "社員一覧（管理）",
                "operationId": "f724a30dd6f2c3451a01c1e997c650f1",
                "parameters": [
                    {
                        "name": "department",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "社員一覧"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin/Members"
                ],
                "summary": "社員作成",
                "operationId": "4e59a8154b7b45478318640eeb100834",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "作成成功"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/members/{member}": {
            "put": {
                "tags": [
                    "Admin/Members"
                ],
                "summary": "社員更新",
                "operationId": "6e7568ef1bda07b0e737daf74323fd96",
                "parameters": [
                    {
                        "name": "member",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "更新成功"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin/Members"
                ],
                "summary": "社員削除",
                "operationId": "47f93340f0d72839a3bf45df5b5a6dbb",
                "parameters": [
                    {
                        "name": "member",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "削除成功"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Sanctum Token",
                "scheme": "bearer"
            }
        }
    },
...
}